+++ features/steps/step_purchase.py
@@ -61,6 +61,7 @@ def reserve_tickets_for_this_event(context):
 
     assert reservation
     assert reservation.trolley.transaction_uuid
+    context.reservation = reservation
     context.transaction_uuid = reservation.trolley.transaction_uuid
 
 
@@ -107,6 +108,17 @@ def invalid_credit_card_details(context):
         expiry_month=10,
         expiry_year=1,
     )
+
+
+@given(u'I have already purchased tickets')
+def given_I_have_already_purchased_tickets(context):
+    context.execute_steps(u"""
+        Given my account is set up to allow a user to buy on credit
+        And an event with availability
+        And I have reserved tickets for my customer for this event
+        And my user has provided valid customer information
+        When I purchase the tickets
+        Then the purchase is succesful""")
 
 
 @when(u'I purchase the tickets')