+++ pyticketswitch/__init__.py
@@ -1,4 +1,4 @@
 from pyticketswitch.client import Client  # NOQA
 
 
-__version__ = "2.7.0"
+__version__ = "2.7.2"